backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2

# Auth
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# AI / RAG
langchain==0.3.1
langchain-community==0.3.1
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
supabase>=2.0.0
pypdf==5.0.1
